Update patch set 10
Patch Set 10:

The gate failure is caused by a requirements update (tosca-parser >= 7.0) in the latest upstream commit for heat-translator.

We were not pinning heat-translator to any commit, but https://review.rdoproject.org/r/4176 is. Once that is merged and a new heat-translator rpm is built, we should be good to merge. We'll need to update to a newer tosca-parser rpm in the deps repo, but that can be done later.

Patch-set: 10