Update patch set 2
Patch Set 2:

Thanks for fixing this. Actually I found this broke the beaker job for puppet-tacker but had not yet took time to fix it by myself.

FYI. We need the same fix in stable/ussuri as well since the patch which added the new dependency was merged into Tacker during ussuri cycle.

Patch-set: 2